Moved route tests under single describe for better organization
